requirements.txt
# Production Requirements for PSLR Platform v3.0
# Python 3.10+

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.0

# LLM API Clients
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0

# Optional (for async)
celery==5.3.4
